Implemented: Created route and function to get all valid filters - categories, colors and sizes
